+++ src/wifi/model/eht/eht-frame-exchange-manager.h
@@ -247,6 +247,9 @@ class EhtFrameExchangeManager : public HeFrameExchangeManager
      */
     bool UnblockEmlsrLinksIfAllowed(Mac48Address address, bool checkThisLink);
 
+    bool m_earlyTxopEndDetect; ///< whether the Duration/ID value of the frame being transmitted
+                               ///< or received can be used to early detect an ongoing TXOP end
+
   private:
     /**
      * @return whether the received ICF must be dropped because we are unable to process it